requirements.txt
pandas>=2.0
openpyxl>=3.1
pydicom>=2.4
Pillow>=10.0
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
numpy>=1.24
